# internal/dataapi/api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "v1",
    "title": "FerretDB Data API",
    "description": "A truly Open Source alternative to MongoDB Atlas Data API."
  },
  "servers": [
    {
      "description": "Global App",
      "url": "https://data.mongodb-api.com/app/{appId}/endpoint/data/v1",
      "variables": {
        "appId": {
          "description": "Your App's Client App ID.",
          "default": "your-app-id"
        }
      }
    },
    {
      "description": "Local App",
      "url": "https://{region}.data.mongodb-api.com/app/{appId}/endpoint/data/v1",
      "variables": {
        "appId": {
          "description": "Your App's Client App ID.",
          "default": "your-app-id"
        },
        "region": {
          "description": "The sub-domain for the cloud region where your App is\ndeployed. Uses the format `<region>.<cloud>`. For example,\n`us-east-1.aws` or `eastus2.azure`. To learn more about\ndeployment regions, see [Deployment Models &\nRegions](https://www.mongodb.com/docs/atlas/app-services/apps/deployment-models-and-regions/).\n",
          "default": "us-east-1.aws",
          "enum": [
            "us-east-1.aws",
            "us-east-2.aws",
            "us-west-2.aws",
            "eu-central-1.aws",
            "eu-west-1.aws",
            "eu-west-2.aws",
            "ap-southeast-1.aws",
            "ap-southeast-2.aws",
            "ap-south-1.aws",
            "sa-east-1.aws",
            "eastus2.azure",
...
          ]
        }
...
  "security": [
    {
      "HttpAuth": []
    }
  ],
  "paths": {
...
    "/action/deleteOne": {
      "post": {
        "operationId": "deleteOne",
        "summary": "Delete One Document",
        "description": "Delete a single document that matches a query.",
        "x-codeSamples": [
          {
            "lang": "cURL",
            "label": "cURL",
            "source": "curl -s \"https://data.mongodb-api.com/app/$CLIENT_APP_ID/endpoint/data/v1/action/deleteOne\" \\\n  -X POST \\\n  -H \"apiKey: $API_KEY\" \\\n  -H 'Content-Type: application/ejson' \\\n  -H \"Accept: application/json\" \\\n  -d '{\n    \"dataSource\": \"mongodb-atlas\",\n    \"database\": \"learn-data-api\",\n    \"collection\": \"tasks\",\n    \"filter\": {\n      \"_id\": { \"$oid\": \"64224f3cd79f54ad342dd9b2\" }\n    }\n  }'\n"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/DeleteRequestBody"
                  }
                ],
                "example": {
                  "dataSource": "mongodb-atlas",
                  "database": "todo",
                  "collection": "tasks",
                  "filter": {
                    "text": "Do the dishes"
                  }
                }
              }
            },
            "application/ejson": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/DeleteRequestBody"
                  }
                ],
                "example": {
                  "dataSource": "mongodb-atlas",
                  "database": "todo",
                  "collection": "tasks",
                  "filter": {
                    "text": "Do the dishes"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Deleted",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/DeleteResponseBody"
                    },
                    {
                      "properties": {
                        "deletedCount": {
                          "type": "number"
                        }
                      }
                    }
                  ],
                  "example": {
                    "deletedCount": 1
                  }
                }
              },
              "application/ejson": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/DeleteResponseBody"
                    },
                    {
                      "properties": {
                        "deletedCount": {
                          "type": "object"
                        }
                      }
                    }
                  ],
                  "example": {
                    "deletedCount": {
                      "$numberInt": 1
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "description": "Unauthorized",
            "$ref": "#/components/responses/UnauthorizedRequestError"
          }
        }
      }
    },
...
    "/action/aggregate": {
      "post": {
        "operationId": "aggregate",
        "summary": "Aggregate Documents",
        "description": "Run an aggregation pipeline.",
        "x-codeSamples": [
          {
            "lang": "cURL",
            "label": "cURL",
            "source": "curl -s \"https://data.mongodb-api.com/app/$CLIENT_APP_ID/endpoint/data/v1/action/aggregate\" \\\n  -X POST \\\n  -H \"apiKey: $API_KEY\" \\\n  -H 'Content-Type: application/ejson' \\\n  -H \"Accept: application/json\" \\\n  -d '{\n    \"dataSource\": \"mongodb-atlas\",\n    \"database\": \"learn-data-api\",\n    \"collection\": \"tasks\",\n    \"pipeline\": [\n      {\n        \"$match\": { \"status\": \"complete\" }\n      },\n      {\n        \"$group\": {\n          \"_id\": \"$status\",\n          \"count\": { \"$sum\": 1 },\n          \"tasks\": { \"$push\": \"$text\" }\n        }\n      },\n      {\n        \"$sort\": { \"count\": -1 }\n      }\n    ]\n  }'\n"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/AggregateRequestBody"
                  }
                ],
                "example": [
                  {
                    "$groupBy": {
                      "_id": "$status",
                      "count": {
                        "$sum": 1
                      },
                      "tasks": {
                        "$push": "$text"
                      }
                    }
                  },
                  {
                    "$sort": {
                      "count": 1
                    }
                  }
                ]
              }
            },
            "application/ejson": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/AggregateRequestBody"
                  }
                ],
                "example": [
                  {
                    "$groupBy": {
                      "_id": "$status",
                      "count": {
                        "$sum": 1
                      },
                      "tasks": {
                        "$push": "$text"
                      }
                    }
                  },
                  {
                    "$sort": {
                      "count": 1
                    }
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/AggregateResponseBody"
                    }
                  ],
                  "example": [
                    {
                      "$groupBy": {
                        "_id": "$status",
                        "count": {
                          "$sum": 1
                        },
                        "tasks": {
                          "$push": "$text"
                        }
                      }
                    },
                    {
                      "$sort": {
                        "count": 1
                      }
                    }
                  ]
                }
              },
              "application/ejson": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/AggregateResponseBody"
                    }
                  ],
                  "example": [
                    {
                      "$groupBy": {
                        "_id": "$status",
                        "count": {
                          "$sum": 1
                        },
                        "tasks": {
                          "$push": "$text"
                        }
                      }
                    },
                    {
                      "$sort": {
                        "count": 1
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "description": "Unauthorized",
            "$ref": "#/components/responses/UnauthorizedRequestError"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "HttpAuth": {
        "type": "http",
        "scheme": "basic"
      },
      "AccessToken": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "A [user access\ntoken](https://www.mongodb.com/docs/atlas/app-services/users/sessions/)\nspecified as a Bearer token in the Authorization header.\n\n**Example:** `Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...`"
      },
      "Email": {
        "type": "apiKey",
        "in": "header",
        "name": "email",
        "description": "A registered email/password user's email address. Must be used\ntogether with the ``password`` header.\n\n**Example:** `email: someone@example.com`"
      },
      "Password": {
        "type": "apiKey",
        "in": "header",
        "name": "password",
        "description": "A registered email/password user's password. Must be used\ntogether with the ``email`` header.\n\n**Example:** `password: Pa55w0rd!`"
      },
      "ApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "apiKey",
        "description": "A server or user API key.\n\n**Example:** `apiKey: A1g70fN7oKMNCnDRNeyou20Wd24rD6QNmXWYMPwz0SnDRQRQSYwGPdC3y44aJN3H`"
      },
      "CustomJwt": {
        "type": "apiKey",
        "in": "header",
        "name": "jwtTokenString",
        "description": "A custom JSON web token string that the App is configured to support.\n\n**Example:** `jwtTokenString: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJhdWQiOiJteWFwcC1hYmNkZSIsInN1YiI6IjY0MzZlMDZiZGU1YzQ2ZGQ5MTQwZWFlOSIsIm5hbWUiOiJKb2huIERvZSIsImlhdCI6MTUxNjIzOTAyMn0.Y_6JR4ocrqoCEY3yOZlEqhxtAsSpvlzrXc9uvTSJA0k`"
      }
    },
    "schemas": {
      "Namespace": {
        "type": "object",
        "required": [
          "dataSource",
          "database",
          "collection"
        ],
        "properties": {
          "dataSource": {
            "type": "string",
            "description": "The name of a linked MongoDB Atlas data source. This is\ncommonly `\"mongodb-atlas\"` though it may be different in\nyour App if you chose a different name when you created the\ndata source.\n"
          },
          "database": {
            "type": "string",
            "description": "The name of a database in the specified data source."
          },
          "collection": {
            "type": "string",
            "description": "The name of a collection in the specified database."
          }
        }
      },
      "Filter": {
        "type": "object",
        "properties": {
          "filter": {
            "type": "object",
            "x-go-type": "json.RawMessage",
            "description": "A MongoDB query filter that matches documents. For a list of all query operators that the Data API supports, see [Query Operators](https://www.mongodb.com/docs/atlas/app-services/mongodb/crud-and-aggregation-apis/#query-operators)."
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "x-go-type-skip-optional-pointer": true,
            "description": "A message that describes the error."
          },
          "error_code": {
            "type": "string",
            "x-go-type-skip-optional-pointer": true,
            "description": "The error type."
          },
          "link": {
            "type": "string",
            "x-go-type-skip-optional-pointer": true,
            "description": "A link to a [log entry](https://www.mongodb.com/docs/atlas/app-services/logs/endpoint/) for the failed operation."
          }
        }
      },
      "ErrorNoAuthenticationSpecified": {
        "description": "Indicates that the request did not include any authentication\ncredentials.",
        "allOf": [
          {
            "$ref": "#/components/schemas/Error"
          },
          {
            "properties": {
              "error": {
                "const": "no authentication methods were specified"
              },
              "error_code": {
                "const": "InvalidParameter"
              }
            }
          }
        ]
      },
      "ErrorMissingAuthenticationParameter": {
        "description": "Indicates that an authentication parameter is missing from the\nrequest body or headers. For example, the request might include\nan `email` header but no corresponding `password` header.",
        "allOf": [
          {
            "$ref": "#/components/schemas/Error"
          },
          {
            "properties": {
              "error": {
                "const": "must specify some form of authentication (either email+password, api-key, or jwt) in the request header or body"
              },
              "error_code": {
                "const": "MissingParameter"
              }
            }
          }
        ]
      },
      "ErrorUserNotFound": {
        "description": "Indicates that no user matched the provided authentication\ncredentials.",
        "allOf": [
          {
            "$ref": "#/components/schemas/Error"
          },
          {
            "properties": {
              "error": {
                "const": "invalid session: error finding user for endpoint"
              },
              "error_code": {
                "const": "InvalidSession"
              }
            }
          }
        ]
      },
...
      "DeleteRequestBody": {
        "title": "DeleteRequestBody",
        "allOf": [
          {
            "$ref": "#/components/schemas/Namespace"
          },
          {
            "$ref": "#/components/schemas/Filter"
          }
        ]
      },
      "DeleteResponseBody": {
        "title": "DeleteResponseBody",
        "type": "object",
        "required": [
          "deletedCount"
        ],
        "properties": {
          "deletedCount": {
            "description": "The number of documents that were deleted."
          }
        }
      },
      "AggregateRequestBody": {
        "title": "AggregateRequestBody",
        "required": [
          "pipeline"
        ],
        "allOf": [
          {
            "$ref": "#/components/schemas/Namespace"
          },
          {
            "properties": {
              "pipeline": {
                "type": "array",
                "x-go-type": "json.RawMessage",
                "description": "An array of aggregation stages.",
                "items": {
                  "type": "object",
                  "description": "A MongoDB aggregation stage. For a list of all aggregation stages that the Data API supports, see [Aggregation Pipeline Stage Availability](https://www.mongodb.com/docs/atlas/app-services/mongodb/crud-and-aggregation-apis/#aggregation)."
                }
              }
            }
          }
        ]
      },
      "AggregateResponseBody": {
        "title": "AggregateResponseBody",
        "required": [
          "documents"
        ],
        "properties": {
          "documents": {
            "description": "An array that contains the result set of the aggregation.",
            "type": "array",
            "x-go-type": "json.RawMessage",
            "items": {
              "type": "object",
              "description": "A document included in the result set of the aggregation."
            }
          }
        }
      }
    },
    "responses": {
      "BadRequestError": {
        "description": "The request was malformed or incomplete.",
        "content": {
          "application/json": {
            "schema": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/ErrorNoAuthenticationSpecified"
                },
                {
                  "$ref": "#/components/schemas/ErrorMissingAuthenticationParameter"
                }
              ]
            }
          }
        }
      },
      "UnauthorizedRequestError": {
        "description": "The authenticated user does not have permission to access this endpoint.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorUserNotFound"
            }
          }
        }
      }
    }
  }
}